+++ var/dump/m2_item_variation_chars.strc.sql
@@ -1,6 +1,6 @@
 CREATE TABLE `m2_item_variation_chars` (
-  `id` smallint(5) unsigned NOT NULL AUTO_INCREMENT,
-  `m2_id` smallint(5) unsigned NOT NULL DEFAULT '0',
+  `id` int(11) unsigned NOT NULL AUTO_INCREMENT,
+  `m2_id` int(11) unsigned NOT NULL DEFAULT '0',
   `parent_id` int(11) unsigned NOT NULL,
   `type_value` tinyint(1) unsigned NOT NULL DEFAULT '0',
   `type_id` int(10) unsigned NOT NULL DEFAULT '0',
